+++ src/app/i18n.module.ts
@@ -1,0 +1,31 @@
+import {NgModule} from "@angular/core";
+import {TranslateLoader, TranslateModule, TranslateService} from "@ngx-translate/core";
+import {HttpClient, HttpClientModule} from "@angular/common/http";
+import {TranslateHttpLoader} from "@ngx-translate/http-loader";
+
+export function HttpLoaderFactory(http: HttpClient) {
+  return new TranslateHttpLoader(http);
+}
+
+@NgModule({
+  imports: [
+    HttpClientModule,
+    TranslateModule.forRoot({
+      defaultLanguage: 'en',
+      loader: {
+        provide: TranslateLoader,
+        useFactory: HttpLoaderFactory,
+        deps: [HttpClient]
+      }
+    }),
+  ],
+  exports: [TranslateModule],
+})
+export class I18nModule {
+  constructor(
+    service: TranslateService,
+  ) {
+    service.setDefaultLang("en");
+    service.use("en");
+  }
+}
